@@ -3,9 +3,9 @@
   "dependencies": {
     "@hapi/joi": "17.1.1",
     "@wireapp/certificate-check": "0.1.12",
-    "@wireapp/commons": "3.6.0",
-    "@wireapp/react-ui-kit": "7.32.2",
-    "@wireapp/webapp-events": "0.3.0",
+    "@wireapp/commons": "3.6.3",
+    "@wireapp/react-ui-kit": "7.36.2",
+    "@wireapp/webapp-events": "0.5.0",
     "auto-launch": "5.0.5",
     "axios": "0.19.2",
     "content-type": "1.0.4",
@@ -25,7 +25,7 @@
     "raygun": "0.12.1",
     "react": "16.13.1",
     "react-dom": "16.13.1",
-    "react-redux": "7.2.0",
+    "react-redux": "7.2.1",
     "redux": "4.0.5",
     "redux-logger": "3.0.6",
     "redux-thunk": "2.3.0",
@@ -33,41 +33,41 @@
   },
   "description": "The most secure collaboration platform.",
   "devDependencies": {
-    "@babel/core": "7.10.4",
+    "@babel/core": "7.11.0",
     "@babel/plugin-proposal-class-properties": "7.10.4",
-    "@babel/plugin-proposal-optional-chaining": "7.10.4",
-    "@babel/preset-env": "7.10.4",
+    "@babel/plugin-proposal-optional-chaining": "7.11.0",
+    "@babel/preset-env": "7.11.0",
     "@babel/preset-react": "7.10.4",
     "@babel/preset-typescript": "7.10.4",
-    "@babel/register": "7.10.4",
+    "@babel/register": "7.10.5",
     "@types/auto-launch": "5.0.1",
     "@types/adm-zip": "0.4.33",
     "@types/content-type": "1.1.3",
     "@types/fs-extra": "9.0.1",
     "@types/is-ci": "2.0.0",
-    "@types/lodash": "4.14.157",
+    "@types/lodash": "4.14.158",
     "@types/minimist": "1.2.0",
-    "@types/mocha": "7.0.2",
+    "@types/mocha": "8.0.1",
     "@types/node": "~12",
     "@types/open-graph": "0.2.0",
     "@types/sinon": "9.0.4",
     "@typescript-eslint/eslint-plugin": "2.34.0",
     "@typescript-eslint/parser": "2.34.0",
-    "@wireapp/copy-config": "1.0.13",
+    "@wireapp/copy-config": "1.1.0",
     "@wireapp/eslint-config": "1.7.0",
     "@wireapp/prettier-config": "0.3.0",
     "adm-zip": "0.4.16",
-    "aws-sdk": "2.713.0",
+    "aws-sdk": "2.725.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "10.1.0",
-    "babel-jest": "26.1.0",
+    "babel-jest": "26.2.2",
     "babel-loader": "8.1.0",
     "babel-plugin-istanbul": "6.0.0",
-    "commander": "5.1.0",
+    "commander": "6.0.0",
     "core-js": "2.6.11",
     "cross-env": "7.0.2",
     "cspell": "4.0.63",
-    "css-loader": "3.6.0",
+    "css-loader": "4.2.0",
     "dotenv": "8.2.0",
     "electron": "10.0.0-beta.10",
     "electron-builder": "20.44.4",
@@ -79,22 +79,22 @@
     "eslint-config-prettier": "6.11.0",
     "eslint-plugin-import": "2.22.0",
     "eslint-plugin-jasmine": "4.1.1",
-    "eslint-plugin-jsdoc": "29.1.3",
+    "eslint-plugin-jsdoc": "30.1.0",
     "eslint-plugin-no-unsanitized": "3.1.2",
     "eslint-plugin-prettier": "3.1.4",
-    "eslint-plugin-react": "7.20.3",
-    "eslint-plugin-react-hooks": "4.0.7",
+    "eslint-plugin-react": "7.20.5",
+    "eslint-plugin-react-hooks": "4.0.8",
     "eslint-plugin-simple-import-sort": "5.0.3",
     "eslint-plugin-sort-keys-fix": "1.1.1",
-    "eslint-plugin-typescript-sort-keys": "1.2.0",
+    "eslint-plugin-typescript-sort-keys": "1.3.0",
     "form-data": "3.0.0",
     "globby": "11.0.1",
     "husky": "4.2.5",
     "is-ci": "2.0.0",
-    "jest": "26.1.0",
+    "jest": "26.2.2",
     "lint-staged": "10.2.11",
-    "mocha": "8.0.1",
-    "nock": "13.0.2",
+    "mocha": "8.1.0",
+    "nock": "13.0.3",
     "nyc": "15.1.0",
     "prettier": "2.0.5",
     "react": "16.13.1",
@@ -103,8 +103,8 @@
     "sort-json": "2.0.0",
     "style-loader": "1.2.1",
     "ts-node": "8.10.2",
-    "typescript": "3.9.6",
-    "webpack": "4.43.0",
+    "typescript": "3.9.7",
+    "webpack": "4.44.1",
     "webpack-cli": "3.3.12"
   },
   "homepage": "https://wire.com",
