@@ -37,9 +37,10 @@ import * as logdown from 'logdown';
 import * as minimist from 'minimist';
 import * as path from 'path';
 import {URL} from 'url';
-import WindowStateKeeper = require('electron-window-state');
+import windowStateKeeper = require('electron-window-state');
 import fileUrl = require('file-url');
 
+import './global';
 import {
   attachTo as attachCertificateVerifyProcManagerTo,
   setCertificateVerifyProc,
@@ -52,7 +53,7 @@ import * as locale from './locale/locale';
 import {ENABLE_LOGGING, getLogger} from './logging/getLogger';
 import {Raygun} from './logging/initRaygun';
 import {getLogFilenames} from './logging/loggerUtils';
-import {menuItem as developerMenu} from './menu/developer';
+import {developerMenu} from './menu/developer';
 import * as systemMenu from './menu/system';
 import {TrayHandler} from './menu/TrayHandler';
 import * as EnvironmentUtil from './runtime/EnvironmentUtil';
@@ -65,7 +66,7 @@ import {SingleSignOn} from './sso/SingleSignOn';
 import {AboutWindow} from './window/AboutWindow';
 import {ProxyPromptWindow} from './window/ProxyPromptWindow';
 import {WindowManager} from './window/WindowManager';
-import {WindowUtil} from './window/WindowUtil';
+import * as WindowUtil from './window/WindowUtil';
 import * as ProxyAuth from './auth/ProxyAuth';
 import {showErrorDialog} from './lib/showDialog';
 
@@ -93,6 +94,7 @@ const BASE_URL = EnvironmentUtil.web.getWebappUrl(argv[config.ARGUMENT.ENV]);
 
 const logger = getLogger(path.basename(__filename));
 const currentLocale = locale.getCurrent();
+const startHidden = Boolean(argv[config.ARGUMENT.STARTUP] || argv[config.ARGUMENT.HIDDEN]);
 
 if (argv[config.ARGUMENT.VERSION]) {
   console.info(config.version);
@@ -136,7 +138,7 @@ Object.entries(config).forEach(([key, value]) => {
 app.setAppUserModelId(`com.squirrel.wire.${config.name.toLowerCase()}`);
 
 // IPC events
-const bindIpcEvents = () => {
+const bindIpcEvents = (): void => {
   ipcMain.on(EVENT_TYPE.ACTION.SAVE_PICTURE, (_event, bytes: Uint8Array, timestamp?: string) => {
     return downloadImage(bytes, timestamp);
   });
@@ -157,14 +159,14 @@ const bindIpcEvents = () => {
   ipcMain.on(EVENT_TYPE.ABOUT.SHOW, AboutWindow.showWindow);
 };
 
-const checkConfigV0FullScreen = (mainWindowState: WindowStateKeeper.State) => {
+const checkConfigV0FullScreen = (mainWindowState: windowStateKeeper.State): void => {
   // if a user still has the old config version 0 and had the window maximized last time
   if (typeof mainWindowState.isMaximized === 'undefined' && isFullScreen === true) {
     main.maximize();
   }
 };
 
-const initWindowStateKeeper = () => {
+const initWindowStateKeeper = (): windowStateKeeper.State => {
   const loadedWindowBounds = settings.restore(SettingsType.WINDOW_BOUNDS, {
     height: WINDOW_SIZE.DEFAULT_HEIGHT,
     width: WINDOW_SIZE.DEFAULT_WIDTH,
@@ -173,13 +175,7 @@ const initWindowStateKeeper = () => {
   // load version 0 full screen setting
   const showInFullScreen = settings.restore(SettingsType.FULL_SCREEN, 'not-set-in-v0');
 
-  const stateKeeperOptions: {
-    defaultHeight: number;
-    defaultWidth: number;
-    fullScreen?: boolean;
-    maximize?: boolean;
-    path: string;
-  } = {
+  const stateKeeperOptions: windowStateKeeper.Options = {
     defaultHeight: loadedWindowBounds.height,
     defaultWidth: loadedWindowBounds.width,
     path: path.join(app.getPath('userData'), 'config'),
@@ -191,11 +187,11 @@ const initWindowStateKeeper = () => {
     isFullScreen = showInFullScreen as boolean;
   }
 
-  return WindowStateKeeper(stateKeeperOptions);
+  return windowStateKeeper(stateKeeperOptions);
 };
 
 // App Windows
-const showMainWindow = async (mainWindowState: WindowStateKeeper.State) => {
+const showMainWindow = async (mainWindowState: windowStateKeeper.State): Promise<void> => {
   const showMenuBar = settings.restore(SettingsType.SHOW_MENU_BAR, true);
 
   const options: BrowserWindowConstructorOptions = {
@@ -230,21 +226,22 @@ const showMainWindow = async (mainWindowState: WindowStateKeeper.State) => {
   attachCertificateVerifyProcManagerTo(main);
   checkConfigV0FullScreen(mainWindowState);
 
-  let webappUrl = `${BASE_URL}${BASE_URL.includes('?') ? '&' : '?'}hl=${currentLocale}`;
+  const webappURL = new URL(BASE_URL);
+  webappURL.searchParams.set('hl', currentLocale);
 
   if (ENABLE_LOGGING) {
-    webappUrl += '&enableLogging=@wireapp/*';
+    webappURL.searchParams.set('enableLogging', '@wireapp/*');
   }
 
   if (customProtocolHandler.hashLocation) {
-    webappUrl += `#${customProtocolHandler.hashLocation}`;
+    webappURL.hash = customProtocolHandler.hashLocation;
   }
 
   if (argv.devtools) {
     main.webContents.openDevTools({mode: 'detach'});
   }
 
-  if (!argv[config.ARGUMENT.STARTUP] && !argv[config.ARGUMENT.HIDDEN]) {
+  if (!startHidden) {
     if (!WindowUtil.isInView(main)) {
       main.center();
     }
@@ -318,17 +315,17 @@ const showMainWindow = async (mainWindowState: WindowStateKeeper.State) => {
     }
   });
 
-  await main.loadURL(`${fileUrl(INDEX_HTML)}?env=${encodeURIComponent(webappUrl)}`);
+  const mainURL = new URL(fileUrl(INDEX_HTML));
+  mainURL.searchParams.set('env', encodeURIComponent(webappURL.href));
+  mainURL.searchParams.set('focus', String(!startHidden));
+
+  await main.loadURL(mainURL.href);
   const wrapperCSSContent = await fs.readFile(WRAPPER_CSS, 'utf8');
   await main.webContents.insertCSS(wrapperCSSContent);
-
-  if (argv[config.ARGUMENT.STARTUP] || argv[config.ARGUMENT.HIDDEN]) {
-    WindowManager.sendActionToPrimaryWindow(EVENT_TYPE.PREFERENCES.SET_HIDDEN);
-  }
 };
 
 // App Events
-const handleAppEvents = () => {
+const handleAppEvents = (): void => {
   app.on('window-all-closed', async () => {
     if (!EnvironmentUtil.platform.IS_MAC_OS) {
       await lifecycle.quit();
@@ -445,7 +442,7 @@ const renameWebViewLogFiles = (): void => {
   }
 };
 
-const addLinuxWorkarounds = () => {
+const addLinuxWorkarounds = (): void => {
   if (EnvironmentUtil.platform.IS_LINUX) {
     // Fix indicator icon on Unity
     // Source: https://bugs.launchpad.net/ubuntu/+bug/1559249
@@ -460,7 +457,7 @@ const addLinuxWorkarounds = () => {
   }
 };
 
-const handlePortableFlags = () => {
+const handlePortableFlags = (): void => {
   if (argv[config.ARGUMENT.USER_DATA_DIR] || argv[config.ARGUMENT.PORTABLE]) {
     const USER_PATH = argv[config.ARGUMENT.USER_DATA_DIR]
       ? path.resolve(argv[config.ARGUMENT.USER_DATA_DIR])
@@ -471,7 +468,7 @@ const handlePortableFlags = () => {
   }
 };
 
-const applyProxySettings = async (authenticatedProxyDetails: URL, webContents: Electron.WebContents) => {
+const applyProxySettings = async (authenticatedProxyDetails: URL, webContents: Electron.WebContents): Promise<void> => {
   const proxyURL = authenticatedProxyDetails.origin.split('://')[1];
   const proxyProtocol = authenticatedProxyDetails.protocol;
   const isSocksProxy = proxyProtocol === 'socks4:' || proxyProtocol === 'socks5:';
@@ -503,7 +500,7 @@ class ElectronWrapperInit {
       frameName: string,
       _disposition: string,
       options: BrowserWindowConstructorOptions,
-    ) => {
+    ): Promise<void> => {
       event.preventDefault();
 
       if (SingleSignOn.isSingleSignOnLoginWindow(frameName)) {
@@ -514,7 +511,7 @@ class ElectronWrapperInit {
       return shell.openExternal(url);
     };
 
-    const willNavigateInWebview = (event: ElectronEvent, url: string, baseUrl: string) => {
+    const willNavigateInWebview = (event: ElectronEvent, url: string, baseUrl: string): void => {
       // Ensure navigation is to an allowed domain
       if (OriginValidator.isMatchingHost(url, baseUrl)) {
         this.logger.log(`Navigating inside webview. URL: ${url}`);
@@ -601,10 +598,10 @@ class ElectronWrapperInit {
             const listenerOnHeadersReceived = (
               details: OnHeadersReceivedListenerDetails,
               callback: (response: HeadersReceivedResponse) => void,
-            ) => {
+            ): void => {
               const responseHeaders = {
                 'Access-Control-Allow-Credentials': ['true'],
-                'Access-Control-Allow-Origin': ['http://localhost:8081'],
+                'Access-Control-Allow-Origin': [EnvironmentUtil.URL_WEBAPP.LOCALHOST],
               };
 
               callback({
