@@ -18,7 +18,8 @@
  */
 
 import {generateUUID} from '../../lib/util';
-import {ActionType, addAccount, deleteAccount, switchAccount, updateAccount, updateAccountBadge} from '../';
+import {ActionType, addAccount, deleteAccount, updateAccount, updateAccountBadge} from '../';
+import {switchAccount} from '../AccountAction';
 
 describe('action creators', () => {
   describe('addAccount', () => {
