@@ -17,17 +17,15 @@
  *
  */
 
-import React, {useEffect, useState, useRef} from 'react';
-import {ContainerSM, Text, H1, Logo, TextLink} from '@wireapp/react-ui-kit';
+import React, {useEffect, useRef, useState} from 'react';
+import {ContainerSM, H1, Logo, Text, TextLink} from '@wireapp/react-ui-kit';
 import {SVGIcon} from '@wireapp/react-ui-kit/dist/Icon/SVGIcon';
 import {connect} from 'react-redux';
-
 import {EVENT_TYPE} from '../../../src/lib/eventType';
 import {WindowUrl} from '../../../src/window/WindowUrl';
 import {
   abortAccountCreation,
   resetIdentity,
-  switchAccount,
   updateAccountBadgeCount,
   updateAccountData,
   updateAccountLifecycle,
@@ -36,6 +34,7 @@ import {getText} from '../lib/locale';
 import {AccountSelector} from '../selector/AccountSelector';
 
 import './Webview.css';
+import {accountAction} from '../actions/AccountAction';
 
 const getEnvironmentUrl = account => {
   const currentLocation = new URL(window.location.href);
@@ -56,10 +55,11 @@ const getEnvironmentUrl = account => {
 
 const Webview = ({
   account,
+  accountIndex,
   onUnreadCountUpdated,
   abortAccountCreation,
   resetIdentity,
-  switchAccount,
+  switchWebview,
   updateAccountData,
   updateAccountLifecycle,
 }) => {
@@ -83,16 +83,22 @@ const Webview = ({
   // https://github.com/electron/electron/issues/14474#issuecomment-425794480
   useEffect(() => {
     const webview = webviewRef.current;
-    const listener = () => {
-      webview.blur();
-      webview.focus();
+    const currentLocation = new URL(window.location.href);
+    const focusParam = currentLocation.searchParams.get('focus');
+
+    const focusWebView = () => {
+      if (focusParam === 'true') {
+        webview.blur();
+        webview.focus();
+      }
     };
+
     if (account.visible && webview) {
-      webview.addEventListener('dom-ready', listener);
+      webview.addEventListener('dom-ready', focusWebView);
     }
     return () => {
       if (webview) {
-        webview.removeEventListener('dom-ready', listener);
+        webview.removeEventListener('dom-ready', focusWebView);
       }
     };
   }, [account, webviewRef]);
@@ -151,7 +157,7 @@ const Webview = ({
         }
 
         case EVENT_TYPE.ACTION.NOTIFICATION_CLICK: {
-          switchAccount(account.id);
+          switchWebview(accountIndex);
           break;
         }
 
@@ -270,10 +276,10 @@ const Webview = ({
   );
 };
 
-export default connect(state => ({accounts: AccountSelector.getAccounts(state)}), {
+export default connect((state, props) => ({accountIndex: AccountSelector.getAccountIndex(state, props.account.id)}), {
   abortAccountCreation,
   resetIdentity,
-  switchAccount,
+  switchWebview: accountAction.switchWebview,
   updateAccountBadgeCount,
   updateAccountData,
   updateAccountLifecycle,
