@@ -19,8 +19,9 @@
 
 /* eslint-disable no-magic-numbers */
 
-import {addAccount, deleteAccount, switchAccount, updateAccount, updateAccountBadge} from '../../actions';
+import {addAccount, deleteAccount, updateAccount, updateAccountBadge} from '../../actions';
 import accountReducer from '../accountReducer';
+import {switchAccount} from '../../actions/AccountAction';
 
 describe('accounts reducer', () => {
   it('should return the initial state with one account', () => {
