@@ -19,8 +19,8 @@
 
 import {DateUtil} from '@wireapp/commons';
 import {dialog, SaveDialogOptions} from 'electron';
-import * as fs from 'fs-extra';
 import imageType from 'image-type';
+import * as fs from 'fs-extra';
 import * as path from 'path';
 
 import {getLogger} from '../logging/getLogger';
@@ -40,7 +40,7 @@ export const downloadLogs = async (bytes: Uint8Array, timestamp: Date = new Date
   return downloadFile(bytes, filename, options);
 };
 
-export const downloadImage = async (bytes: Uint8Array, timestamp?: string) => {
+export const downloadImage = async (bytes: Uint8Array, timestamp?: string): Promise<void> => {
   const type = imageType(bytes);
   const options: SaveDialogOptions = {};
 
@@ -59,7 +59,7 @@ export const downloadImage = async (bytes: Uint8Array, timestamp?: string) => {
   return downloadFile(bytes, filename, options);
 };
 
-export const downloadFile = async (bytes: Uint8Array, filename: string, options?: SaveDialogOptions) => {
+export const downloadFile = async (bytes: Uint8Array, filename: string, options?: SaveDialogOptions): Promise<void> => {
   try {
     const {filePath: chosenPath} = await dialog.showSaveDialog({defaultPath: filename, ...options});
     if (chosenPath) {
@@ -70,7 +70,7 @@ export const downloadFile = async (bytes: Uint8Array, filename: string, options?
   }
 };
 
-export const suggestFileName = (timestamp?: string) => {
+export const suggestFileName = (timestamp?: string): string => {
   const imageDate = timestamp ? new Date(Number(timestamp)) : new Date();
   const {date: formattedDate, time: formattedTime} = DateUtil.isoFormat(imageDate);
   return `Wire ${formattedDate} at ${formattedTime}`.replace(/:/g, '-');
