@@ -19,10 +19,100 @@
 
 import * as Electron from 'electron';
 
-import {i18nLanguageIdentifier, SupportedI18nLanguage, SupportedI18nLanguageObject} from '../interfaces/';
 import {config} from '../settings/config';
 import {settings} from '../settings/ConfigurationPersistence';
 import {SettingsType} from '../settings/SettingsType';
+
+export type i18nLanguageIdentifier =
+  | 'aboutReleases'
+  | 'aboutUpdate'
+  | 'aboutVersion'
+  | 'aboutWebappVersion'
+  | 'certificateVerifyProcManagerRetry'
+  | 'certificateVerifyProcManagerShowDetails'
+  | 'certificateVerifyProcManagerShowDetailsGoBack'
+  | 'certificateVerifyProcManagerShowDetailsSaveCertificate'
+  | 'certificateVerifyProcManagerShowDetailsTextChromium'
+  | 'certificateVerifyProcManagerShowDetailsTextPinning'
+  | 'certificateVerifyProcManagerShowDetailsTitle'
+  | 'certificateVerifyProcManagerWarningBypass'
+  | 'certificateVerifyProcManagerWarningTextChromium'
+  | 'certificateVerifyProcManagerWarningTextPinning'
+  | 'certificateVerifyProcManagerWarningTitle'
+  | 'menuAbout'
+  | 'menuAddPeople'
+  | 'menuAppURL'
+  | 'menuArchive'
+  | 'menuBlock'
+  | 'menuCall'
+  | 'menuClose'
+  | 'menuConversation'
+  | 'menuCopy'
+  | 'menuCut'
+  | 'menuDelete'
+  | 'menuDownloadDebugLogs'
+  | 'menuEdit'
+  | 'menuEnableSpellChecking'
+  | 'menuFullScreen'
+  | 'menuHelp'
+  | 'menuHideApp'
+  | 'menuHideOthers'
+  | 'menuLeave'
+  | 'menuLegal'
+  | 'menuLicense'
+  | 'menuLocale'
+  | 'menuMinimize'
+  | 'menuNextConversation'
+  | 'menuNoSuggestions'
+  | 'menuPaste'
+  | 'menuPeople'
+  | 'menuPing'
+  | 'menuPreferences'
+  | 'menuPreviousConversation'
+  | 'menuPrivacy'
+  | 'menuQuit'
+  | 'menuRedo'
+  | 'menuSavePictureAs'
+  | 'menuSelectAll'
+  | 'menuServices'
+  | 'menuSettings'
+  | 'menuShowAll'
+  | 'menuShowHide'
+  | 'menuSignOut'
+  | 'menuStart'
+  | 'menuStartup'
+  | 'menuSupport'
+  | 'menuUnarchive'
+  | 'menuUndo'
+  | 'menuVideoCall'
+  | 'menuView'
+  | 'menuWindow'
+  | 'proxyPromptCancel'
+  | 'proxyPromptHeadline'
+  | 'proxyPromptOk'
+  | 'proxyPromptPassword'
+  | 'proxyPromptTitle'
+  | 'proxyPromptUsername'
+  | 'restartLater'
+  | 'restartLocale'
+  | 'restartNeeded'
+  | 'restartNow'
+  | 'trayOpen'
+  | 'trayQuit'
+  | 'unreadMessages'
+  | 'wrapperAddAccount'
+  | 'wrapperAddAccountErrorMessagePlural'
+  | 'wrapperAddAccountErrorMessageSingular'
+  | 'wrapperAddAccountErrorTitlePlural'
+  | 'wrapperAddAccountErrorTitleSingular'
+  | 'wrapperCreateTeam'
+  | 'wrapperLogOut'
+  | 'wrapperManageTeam'
+  | 'wrapperRemoveAccount';
+
+export type i18nStrings = Record<i18nLanguageIdentifier, string>;
+export type SupportedI18nLanguage = keyof typeof SUPPORTED_LANGUAGES;
+export type SupportedI18nLanguageObject = Record<SupportedI18nLanguage, i18nStrings>;
 
 const cs_CZ = require('../../locale/cs-CZ');
 const da_DK = require('../../locale/da-DK');
